prospect-agent/pyproject.toml
[tool.poetry]
name = "prospect-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
